fix(kkem): new redesign for interest group muid section
