+++ src/modules/KKEM/components/Auth.module.css
@@ -1,5 +1,6 @@
 .container {
-    padding-top: 2rem;
+    width: 100%;
+    max-width: 40rem;
 }
 .heading {
     font-size: 2rem;
@@ -31,29 +32,21 @@
 .form {
     display: flex;
     height: 5rem;
-    width: fit-content;
     position: relative;
     width: 100%;
-    max-width: 35rem;
     margin-top: 1rem;
     isolation: isolate;
+    border-radius: 1.6rem;
+    background: #fff;
+    box-shadow: 0px 0px 41px 0px rgba(0, 0, 0, 0.1);
 }
-.form::before {
-    content: "";
-    width: 5rem;
-    height: 6rem;
-    background: url("../assets/Star.svg") no-repeat;
-    position: absolute;
-    left: -2.5rem;
-    z-index: -1;
-}
+
 .form > input[type="text"] {
     background-color: #fff;
     width: 100%;
     margin-top: 0;
     border: none;
     border-radius: 1.5rem;
-    padding-left: 2rem;
 }
 
 @media screen and (max-width: 600px) {
