+++ src/modules/KKEM/components/MulearnAbout.module.css
@@ -10,22 +10,18 @@
     font-size: 5rem;
     font-weight: 600;
     line-height: 111.5%;
-    margin-bottom: 5rem;
+    margin-bottom: 3rem;
 }
 
-.title::after {
-    content: "";
-    display: block;
-    width: calc(2.3rem * 14);
-    height: 0.6rem;
-    background: #67de82;
-    margin-top: -1rem;
-    border-radius: 0.5rem;
-    max-width: 100%;
-}
 .description {
     margin-top: 2rem;
     color: #293338;
+    font-weight: 500;
+    font-size: 1.8rem;
+    line-height: 127.5%;
+    max-width: 60vmax;
+}
+.dc2 {
     font-weight: 500;
     font-size: 1.8rem;
     line-height: 127.5%;
@@ -37,7 +33,35 @@
     text-decoration: none;
 }
 .image {
+    width: 30rem;
     animation: upDown 2s ease-in-out infinite;
+}
+.muidSection {
+    display: flex;
+    justify-content: space-between;
+}
+.imgc {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
+.muidLink {
+    color: #2e85fe;
+    font-size: 1.8rem;
+    font-weight: 500;
+    text-decoration-line: underline;
+    line-height: 5rem;
+}
+
+@media screen and (max-width: 1200px) {
+    .muidSection {
+        flex-direction: column;
+    }
+    .muidLink {
+        font-size: 1.2rem;
+    }
 }
 
 @keyframes upDown {
@@ -75,10 +99,6 @@
     .title {
         font-size: 2.5rem;
         margin-bottom: 2rem;
-    }
-    .title::after {
-        width: calc(2rem * 8);
-        margin-top: -0.5rem;
     }
     .description {
         font-size: 1.5rem;
