+++ src/modules/KKEM/components/Navbar.module.css
@@ -1,7 +1,6 @@
 .navbar {
     background-color: #ffffff;
-    padding: 2rem;
-    height: 7rem;
+    padding: 2rem 0;
     display: flex;
     flex-direction: column;
     justify-content: center;
@@ -12,18 +11,13 @@
     justify-content: space-between;
     background-color: #fff;
     padding: 1rem 2rem;
-    gap: 8rem;
+    gap: 3rem;
     border-radius: 1.2rem;
-    top: 2rem;
-    z-index: 2;
     width: 100%;
     max-width: 800px;
     perspective: 1000px;
 }
-.logos {
-    display: flex;
-    gap: 0.5rem;
-}
+
 .container > a {
     color: #000;
     text-decoration: none;
