+++ src/modules/KKEM/components/Navbar.tsx
@@ -1,9 +1,4 @@
-import { useCallback, useEffect, useState } from "react";
-import kkemlogo from "../assets/kkemlogo.png";
-import XSvg from "../assets/X.svg";
-import mulearnLogo from "../assets/µ.svg";
 import styles from "./Navbar.module.css";
-import { RiMenu4Line, RiCloseLine } from "react-icons/ri";
 export default function Navbar() {
     return (
         <nav className={styles.navbar}>
